Implement support for CSL-JSON
The CSL-JSON is now used as a primaty format (as it supports more
properties and stores them in a more structured way) with a possible
conversion to the BibLaTeX. This requires provision of mappings in both
directions in the config.

index.php now supports content type negotation over the Accept header
and the "format" request parameter.

Overrides can be passed both as BibLaTeX and CSL-JSON.